refactor: Adjust bottom navigation bar visibility
This commit updates the condition for displaying the `BottomNavigationBar` in `MainActivity`.

The visibility logic is changed from showing the bar only on the `Home` screen (`configurationState is ConfigChildComponent.Home`) to showing it on all screens except the `Detail` screen (`configurationState !is ConfigChildComponent.Detail`). This ensures the navigation bar is present on more screens, such as the `Favorite` and `News` tabs.